+++ NEWS.md
@@ -31,6 +31,10 @@ You can find a short guide to  astsa scripts right here at [**FUN WITH ASTSA**](
 
 #### v2.0.1: small updates after v2.0:
 
+- `test.linear` gets a title (`main`) control
+
+- fix to `Kfilter(... , version=2)`
+
 - `mvspec` gets the gris-gris option, enhancing the voodoo grammar of astsa
 
 - made `acf1` more flexible